+++ src/commands/createNewProject/ScriptProjectCreatorBase.ts
@@ -5,7 +5,7 @@
 
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { gitignoreFileName, hostFileName, localSettingsFileName, ProjectRuntime, TemplateFilter } from '../../constants';
+import { gitignoreFileName, hostFileName, localSettingsFileName, ProjectRuntime, proxiesFileName, TemplateFilter } from '../../constants';
 import { ILocalAppSettings } from '../../LocalAppSettings';
 import { confirmOverwriteFile } from "../../utils/fs";
 import * as fsUtil from '../../utils/fs';
@@ -70,7 +70,9 @@ export class ScriptProjectCreatorBase extends ProjectCreatorBase {
     public async addNonVSCodeFiles(): Promise<void> {
         const hostJsonPath: string = path.join(this.functionAppPath, hostFileName);
         if (await confirmOverwriteFile(hostJsonPath, this.ui)) {
-            const hostJson: {} = {};
+            const hostJson: {} = {
+                version: '2.0'
+            };
             await fsUtil.writeFormattedJson(hostJsonPath, hostJson);
         }
 
@@ -91,8 +93,6 @@ export class ScriptProjectCreatorBase extends ProjectCreatorBase {
             await fsUtil.writeFormattedJson(localSettingsJsonPath, localSettingsJson);
         }
 
-        // Temporarily disabling due to https://github.com/Azure/azure-functions-core-tools/issues/562
-        /*
         const proxiesJsonPath: string = path.join(this.functionAppPath, proxiesFileName);
         if (await confirmOverwriteFile(proxiesJsonPath, this.ui)) {
             const proxiesJson: {} = {
@@ -102,7 +102,6 @@ export class ScriptProjectCreatorBase extends ProjectCreatorBase {
             };
             await fsUtil.writeFormattedJson(proxiesJsonPath, proxiesJson);
         }
-        */
 
         const gitignorePath: string = path.join(this.functionAppPath, gitignoreFileName);
         if (await confirmOverwriteFile(gitignorePath, this.ui)) {