+++ src/tree/FunctionAppProvider.ts
@@ -3,18 +3,18 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-// tslint:disable-next-line:no-require-imports
-import WebSiteManagementClient = require('azure-arm-website');
+import { WebSiteManagementClient } from 'azure-arm-website';
 import { Site, WebAppCollection } from "azure-arm-website/lib/models";
 import { OutputChannel } from "vscode";
 import { createFunctionApp, SiteClient } from 'vscode-azureappservice';
-import { addExtensionUserAgent, IActionContext, IAzureNode, IAzureTreeItem, IChildProvider } from 'vscode-azureextensionui';
+import { addExtensionUserAgent, IActionContext, IAzureNode, IAzureTreeItem, IChildProvider, parseError } from 'vscode-azureextensionui';
 import { ProjectRuntime, projectRuntimeSetting } from '../constants';
 import { tryGetLocalRuntimeVersion } from '../funcCoreTools/tryGetLocalRuntimeVersion';
 import { localize } from "../localize";
 import { getFuncExtensionSetting } from '../ProjectSettings';
 import { getCliFeedAppSettings } from '../utils/getCliFeedJson';
 import { FunctionAppTreeItem } from "./FunctionAppTreeItem";
+import { InvalidFunctionAppTreeItem } from './InvalidFunctionAppTreeItem';
 
 export class FunctionAppProvider implements IChildProvider {
     public readonly childTypeLabel: string = localize('azFunc.FunctionApp', 'Function App');
@@ -35,33 +35,51 @@ export class FunctionAppProvider implements IChildProvider {
             this._nextLink = undefined;
         }
 
-        const client: WebSiteManagementClient = new WebSiteManagementClient(node.credentials, node.subscriptionId);
+        const client: WebSiteManagementClient = new WebSiteManagementClient(node.credentials, node.subscriptionId, node.environment.resourceManagerEndpointUrl);
         addExtensionUserAgent(client);
-        const webAppCollection: WebAppCollection = this._nextLink === undefined ?
-            await client.webApps.list() :
-            await client.webApps.listNext(this._nextLink);
+        let webAppCollection: WebAppCollection;
+        try {
+            webAppCollection = this._nextLink === undefined ?
+                await client.webApps.list() :
+                await client.webApps.listNext(this._nextLink);
+        } catch (error) {
+            if (parseError(error).errorType.toLowerCase() === 'notfound') {
+                // This error type means the 'Microsoft.Web' provider has not been registered in this subscription
+                // In that case, we know there are no function apps, so we can return an empty array
+                // (The provider will be registered automatically if the user creates a new function app)
+                return [];
+            } else {
+                throw error;
+            }
+        }
 
         this._nextLink = webAppCollection.nextLink;
 
         const treeItems: IAzureTreeItem[] = [];
         for (const site of webAppCollection) {
-            const siteClient: SiteClient = new SiteClient(site, node);
-            if (siteClient.isFunctionApp) {
-                treeItems.push(new FunctionAppTreeItem(siteClient, this._outputChannel));
+            try {
+                const siteClient: SiteClient = new SiteClient(site, node);
+                if (siteClient.isFunctionApp) {
+                    treeItems.push(new FunctionAppTreeItem(siteClient, this._outputChannel));
+                }
+            } catch (error) {
+                if (site.name) {
+                    treeItems.push(new InvalidFunctionAppTreeItem(site.name, error));
+                }
             }
         }
         return treeItems;
     }
 
-    public async createChild(parent: IAzureNode, showCreatingNode: (label: string) => void, actionContext: IActionContext): Promise<IAzureTreeItem> {
+    public async createChild(parent: IAzureNode, showCreatingNode: (label: string) => void, userOptions?: { actionContext: IActionContext, resourceGroup?: string }): Promise<IAzureTreeItem> {
         // Ideally actionContext should always be defined, but there's a bug with the NodePicker. Create a 'fake' actionContext until that bug is fixed
         // https://github.com/Microsoft/vscode-azuretools/issues/120
         // tslint:disable-next-line:strict-boolean-expressions
-        actionContext = actionContext || <IActionContext>{ properties: {}, measurements: {} };
-
+        const actionContext: IActionContext = userOptions ? userOptions.actionContext : <IActionContext>{ properties: {}, measurements: {} };
+        const resourceGroup: string | undefined = userOptions ? userOptions.resourceGroup : undefined;
         const runtime: ProjectRuntime = await getDefaultRuntime(actionContext);
         const appSettings: { [key: string]: string } = await getCliFeedAppSettings(runtime);
-        const site: Site = await createFunctionApp(actionContext, parent.credentials, parent.subscriptionId, parent.subscriptionDisplayName, showCreatingNode, appSettings);
+        const site: Site = await createFunctionApp(actionContext, parent, showCreatingNode, appSettings, resourceGroup);
         return new FunctionAppTreeItem(new SiteClient(site, parent), this._outputChannel);
     }
 }